refactor: stop importing domain modules from ui/state
tokenRates (state layer) reached into ui/domains for the taoData HTTP
client and the bittensor network id. The client is a leaf (core + util
deps only) and moves to @ui/util; the network id constant now lives in
@ui/state/bittensor (which previously inlined the literal) and is
re-exported from its old domains path for existing consumers.
