+++ apps/extension/src/ui/domains/TaoDashboard/subnets/constants.ts
@@ -1,1 +1,2 @@
-export const BITTENSOR_NETWORK_ID = "bittensor"
+// defined at the state layer so that @ui/state modules don't have to reach into @ui/domains for it
+export { BITTENSOR_NETWORK_ID } from "@ui/state/bittensor"
